Increase the size of AID hashes to 64-bit.
Also fixes a bug when hashing lists.

+++ src/OpenDiffix.Core/CommonTypes.fs
@@ -117,7 +117,7 @@ type IDataProvider =
 // Anonymizer types
 // ----------------------------------------------------------------
 
-type AidHash = int32
+type AidHash = uint64
 
 type Threshold =
   {
